+++ src/main/kotlin/hd/junction/hospital/domain/Hospital.kt
@@ -27,5 +27,4 @@ class Hospital(
     @Id
     @GeneratedValue(strategy = IDENTITY)
     val id: Long? = null,
-) {
-}
+)