+++ src/main/kotlin/hd/junction/patient/application/PatientService.kt
@@ -3,13 +3,12 @@ package hd.junction.patient.application
 import hd.junction.codes.infrastructure.CodeRepository
 import hd.junction.common.util.RandomUtils.generateRandomPatientRegistrationNumber
 import hd.junction.common.util.RandomUtils.getUniquePatientRegistrationNumber
-import hd.junction.hospital.domain.Hospital
 import hd.junction.hospital.infrastructure.HospitalRepository
 import hd.junction.patient.domain.Patient
 import hd.junction.patient.dto.request.PatientRequestDto
 import hd.junction.patient.dto.request.PatientSearchRequestDto
-import hd.junction.patient.dto.response.PatientPageResponseDto
 import hd.junction.patient.dto.response.PatientHospitalResponseDto
+import hd.junction.patient.dto.response.PatientPageResponseDto
 import hd.junction.patient.dto.response.PatientVisitResponseDto
 import hd.junction.patient.infrastructure.PatientQuerydslRepository
 import hd.junction.patient.infrastructure.PatientRepository